cli-plugins/manager: move OTEL-related code to separate file
